Added new file for simulation table view, updated index to be about in preparation for changing front page, deleted unused templates
