--- website/apps/home/templates/home/departmentchart.html
@@ -1,57 +1,0 @@
-{% block extrahead %}
-    <!-- Load in jQuery and HighCharts -->
-    <script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/2.2.2/jquery.min.js"></script>
-    <script type="text/javascript" src="http://code.highcharts.com/highcharts.js"></script>
-{% endblock %}
-
-<style type="text/css">
-    body {
-        background-color: #eee;
-        color: #000;
-        font: 12px sans-serif;
-        margin: 20px;
-    }
-    h1{
-        background-color: #e8e8e8;
-    }
-    .chart {
-        padding: 20px;
-    }
-    .returnbtn{
-        float: right;
-        color: #a6e1ec;
-        font-size: 60%;
-    }
-</style>
-
-{% block content %}
-    <h1 align="center">Zika information of {{ department_name }}
-    <a id = "returnbtn" href = '/home'> Home</a>
-    </h1>
-    <div id={{ chartID|safe }} class="chart" ></div>
-{% endblock %}
-
-{% block extrajs %}
-    <!-- Maps the Python template context variables from views.py to the Highchart js variables -->
-    <script>
-        var chart_id = {{ chartID|safe }};
-        var chart = {{ chart|safe }};
-        var title = {{ title|safe }};
-        var xAxis = {{ xAxis|safe }};
-        var yAxis = {{ yAxis|safe }};
-        var series = {{ series|safe }};
-    </script>
-
-    <!-- Highchart js. Variable map shown above -->
-    <script>
-        $(document).ready(function() {
-            $(chart_id).highcharts({
-                chart: chart,
-                title: title,
-                xAxis: xAxis,
-                yAxis: yAxis,
-                series: series
-            });
-        });
-    </script>
-{% endblock %}
