--- website/apps/home/templates/home/egcsv.html
@@ -1,104 +1,0 @@
-{% extends "base.html" %}
-
-{% block extra_css %}
-    <script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/2.2.2/jquery.min.js"></script>
-    <script type="text/javascript" src="/static/examples/d3.v3.js"></script>
-    <script src="{{ STATIC_URL }}js/jquery-2.1.4.min.js"></script>
-    <script src="{{ STATIC_URL }}js/jquery-migrate-1.2.1.min.js"></script>
-    <script src="{{ STATIC_URL }}bootstrap-3.3.5/js/bootstrap.min.js"></script>
-
-<style type="text/css">
-    body {
-        background-color: white;
-        color: #000;
-        font: 12px sans-serif;
-        margin: 20px;
-    }
-
-    .dropbtn {
-        background-color: #4CAF50;
-        color: white;
-        padding: 16px;
-        font-size: 16px;
-        border: none;
-        cursor: pointer;
-    }
-
-    /* Dropdown button on hover & focus */
-    .dropbtn:hover, .dropbtn:focus {
-        background-color: #3e8e41;
-    }
-
-    /* The container <div> - needed to position the dropdown content */
-    .dropdown {
-        position: relative;
-        display: inline-block;
-    }
-
-    /* Dropdown Content (Hidden by Default) */
-    .dropdown-content {
-        display: none;
-        position: absolute;
-        background-color: #f9f9f9;
-        min-width: 160px;
-        box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
-    }
-
-    /* Links inside the dropdown */
-    .dropdown-content a {
-        color: black;
-        padding: 12px 16px;
-        text-decoration: none;
-        display: block;
-    }
-
-    /* Change color of dropdown links on hover */
-    .dropdown-content a:hover {background-color: #f1f1f1}
-
-    /* Show the dropdown menu (use JS to add this class to the .dropdown-content container when the user clicks on the dropdown button) */
-    .show {display:block;}
-</style>
-{% endblock %}
-
-{% block main_content %}
-    <p><a href="{% url 'home.display_simulations' %}">Back</a></p>
-    <p>Select a date to see the map</p>
-    <div class="dropdown">
-        <button onclick="myFunction()" class="dropbtn">Forcasting Dates</button>
-        <div id="myDropdown" class="dropdown-content">
-            {% for item in municipality_code %}
-                <a href="{% url 'home.choropleth_map' inquery_date=item sim_id=sim_id %}" target="iframe1">{{ item }}</a></td>
-            {% endfor %}
-        </div>
-    </div>
-    <div id="nav">
-    <iframe  name="iframe1" id="link" width="100%" height="700" style="border: none"></iframe>
-    </div>
-{% endblock %}
-
-{% block extra_js %}
-    <!-- Dropdown js.  -->
-    <script>
-        /* When the user clicks on the button, toggle between hiding and showing the dropdown content */
-        function myFunction() {
-            document.getElementById("myDropdown").classList.toggle("show");
-        }
-        // Close the dropdown menu if the user clicks outside of it
-        $(document).ready(function(){
-                    window.onclick = function(event) {
-            if (!event.target.matches('.dropbtn')) {
-                var dropdowns = document.getElementsByClassName("dropdown-content");
-                var i;
-                for (i = 0; i < dropdowns.length; i++) {
-                    var openDropdown = dropdowns[i];
-                    if (openDropdown.classList.contains('show')) {
-                        openDropdown.classList.remove('show');
-                }
-            }
-        }
-    }
-        })
-
-
-    </script>
-{% endblock %}
