+++ website/templates/base.html
@@ -95,7 +95,7 @@
         <ul class="nav nav-pills nav-stacked">
             <li><a href="{% url 'home.display_simulations' %}">Browse Map</a></li>
             <li><a href="{% url 'simulation.upload' %}">Upload Simulation</a></li>
-            <li><a href="{% url  'index'%}">About</a></li>
+            <li><a href="{% url  'about'%}">About</a></li>
         </ul>
     </div>
